app_benchmark/requirements.txt
streamlit>=1.32.0
openai>=1.30.0
azure-search-documents>=11.4.0
azure-core>=1.30.0
azure-ai-evaluation==1.3.0
pandas>=2.0.0
plotly>=5.18.0
